# Protobuf serialization perf improvements