… test/Confluent.SchemaRegistry.IntegrationTests/Tests/Person.cs
… test/Confluent.SchemaRegistry.IntegrationTests/Tests/PersonName.cs
… test/Confluent.SchemaRegistry.IntegrationTests/Tests/User.cs
… test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests_Protobuf/Generated/ManyMessages.cs
… test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests_Protobuf/Generated/Nested.cs
… test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests_Protobuf/Generated/RefByAnother.cs
… test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests_Protobuf/Generated/UInt32Value.cs
… test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests_Protobuf/Generated/WithExternalReference.cs
… test/Confluent.SchemaRegistry.Serdes.IntegrationTests/Tests_Protobuf/Generated/WithGoogleRefs.cs
… test/Confluent.SchemaRegistry.Serdes.UnitTests/Generated/Person.cs
… test/Confluent.SchemaRegistry.Serdes.UnitTests/Generated/PersonWithPic.cs
… test/Confluent.SchemaRegistry.Serdes.UnitTests/Generated/UInt32Value.cs