<!--
Suggested PR template: Fill/delete/add sections as needed. Optionally delete any commented block.
-->
What
----
Optimize serialization by replacing MemoryStream with pre-allocated byte buffers. This reduces allocations and improves performance.

Performance (serialization only)
```
| Method |     Mean |   Error |   StdDev |   Gen0 |   Gen1 | Allocated |
|------- |---------:|--------:|---------:|-------:|-------:|----------:|
|    New | 280.6 ns | 7.27 ns | 20.61 ns | 0.0229 |      - |     144 B |
|    Old | 444.9 ns | 8.21 ns |  7.68 ns | 0.8736 | 0.0014 |    5480 B |
```

Checklist
------------------
No changes in public API
Just an optimization, no functional changes.

## Test & Review
- Manual testing
- Full project test suite


@rayokota Could you please take a look?