@Havret 2025-04-25
Sorry, I hadn’t pushed all the regenerated protos earlier. It should be all set now.