@rayokota 2025-04-25
@Havret , why did the Protobuf classes need to be regenerated?  Is this PR incompatible with customers' older generated classes?