@Havret 2025-04-25
If their classes were generated by `protoc` older than five years, then yes, they would need to regenerate them.  
The API I'm using here — [`WriteTo(this IMessage message, Span<byte> output)`](https://github.com/protocolbuffers/protobuf/commit/9039103637cfeba1aca82c8eead94b1479a6612d) — was introduced in that timeframe.

Realistically, they should regenerate their protos anyway, since older versions of the libraries had known CVEs.